% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindle2anki-package.R
\docType{package}
\name{kindle2anki-package}
\alias{kindle2anki}
\alias{kindle2anki-package}
\title{kindle2anki: Get Vocabulary from Kindle to Anki}
\description{
Import vocabulary from Amazon Kindle e-reader and export to 
    format ready for import as Anki flash cards.
}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

}
\keyword{internal}
