% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindle2anki.R
\name{kindle2anki}
\alias{kindle2anki}
\title{Convert Vocabulary from Kindle to Anki}
\usage{
kindle2anki(
  file,
  asin,
  outfile = paste0(asin, ".tsv"),
  prefer_dictionary_usage = TRUE
)
}
\arguments{
\item{file}{A path to a \code{.db} file.}

\item{asin}{Amazon ID (get it by calling \code{\link[=listBooks]{listBooks()}} on \code{file} first.}

\item{outfile}{Output file; should end with \code{.tsv}.}

\item{prefer_dictionary_usage}{Whether to prefer the usage entry from the
book (Kindle database) or from the wordnet database. If this is \code{TRUE} but
there is no wordnet entry, the usage entry from the book will be returned.}
}
\value{

}
\description{
This function converts and augments
a \verb{*.db} Kindle database file (usually \code{vocab.db})
to a \code{.tsv} file ready for import into Anki. It extracts the words
used and for then, for each word class (adjectives, nouns, verbs, adverbs)
scrapes the wordnet API for a set of definitions, synonyms, and usage.
Only the first usage entry is used. The result is a data table with
each word and its class, definition, synonyms, and usage. You will usually
want to call \code{\link[=listBooks]{listBooks()}} on the \code{file} first and get the Amazon ID
to use in the \code{asin} argument to this function.
}
\seealso{
\code{\link[=listBooks]{listBooks()}}
}
