% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBooks.R
\name{listBooks}
\alias{listBooks}
\title{Get Data Table of Book Titles and Amazon IDs}
\usage{
listBooks(file)
}
\arguments{
\item{file}{input file}
}
\value{
A table of book titles and amazon ids.
}
\description{
Get Data Table of Book Titles and Amazon IDs
}
